pyproject.toml
[tool.poetry]
name = "whatapp"
version = "0.1.0"
description = ""
authors = ["Igor Gerasimov <gerasimov.igor.98@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
telethon = "^1.34.0"
openai = "^1.12.0"
langchain = "^0.1.9"
loguru = "^0.7.2"
fast-bitrix24 = "^1.5.16"
ipywidgets = "^8.1.2"
tiktoken = "^0.6.0"
langchain-community = "^0.0.24"
langchain-openai = "^0.0.8"
grpcio = "^1.62.0"
yandexcloud = "^0.260.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
